requirements.txt
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
idna==2.6
lxml==4.1.1
python-telegram-bot==9.0.0
requests==2.18.4
urllib3==1.22
